backend/requirements-dev.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.7
stripe==7.6.0
twilio==8.10.0
celery==5.3.4
redis==5.0.1
django-filter==23.3
django-csv==2.1
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0

# Alternative to WeasyPrint for development
reportlab==4.0.4
# Alternative to Pillow for development
Pillow==10.1.0 